Re-styled control positions
