+++ frontend/src/scenes/session-recordings/player/controller/Seekbar.scss
@@ -1,6 +1,6 @@
 .PlayerSeekbar {
     --bar-height: 4px;
-    --slider-height: 16px;
+    --slider-height: 20px;
     --thumb-size: 15px;
 
     width: 100%;
@@ -115,73 +115,72 @@
             opacity: 1;
         }
     }
+}
 
-    .PlayerSeekbar__ticks {
-        z-index: 1;
+.PlayerSeekbarTicks {
+    z-index: 1;
+    width: 100%;
+    height: 0.75rem;
+    position: relative;
+
+    .PlayerSeekbarTick {
+        --tick-color: var(--muted-alt);
+        --tick-width: 3px;
+
+        cursor: pointer;
         position: absolute;
-        width: 100%;
-        bottom: 100%;
-        height: 0.75rem;
+        height: 100%;
+        top: 0px;
+        transition: transform 150ms ease-in-out;
 
-        .PlayerSeekbarTick {
-            --tick-color: var(--muted-alt);
-            --tick-width: 3px;
+        &--warning {
+            --tick-color: var(--warning);
+        }
 
-            cursor: pointer;
+        &--danger {
+            --tick-color: var(--danger);
+        }
+
+        &--primary {
+            --tick-color: var(--primary);
+        }
+
+        .PlayerSeekbarTick__line {
             position: absolute;
+            width: var(--tick-width);
+            margin-left: calc(var(--tick-width) / 2 * -1);
+            border-radius: 2px;
+            opacity: 0.5;
             height: 100%;
-            top: 0px;
-            transition: transform 150ms ease-in-out;
+            background-color: var(--tick-color);
+        }
 
-            &--warning {
-                --tick-color: var(--warning);
-            }
+        .PlayerSeekbarTick__info {
+            position: absolute;
+            bottom: 100%;
+            background-color: var(--tick-color);
+            padding: 0.25rem 0.5rem;
+            font-size: 0.8rem;
+            border-radius: 0.25rem;
+            transform: translateX(-50%);
+            white-space: nowrap;
+            color: #fff;
+            margin-bottom: 0.25rem;
+            opacity: 0;
+            visibility: hidden;
+            transition: opacity 150ms ease-in-out;
+            max-width: 20rem;
+        }
 
-            &--danger {
-                --tick-color: var(--danger);
-            }
-
-            &--primary {
-                --tick-color: var(--primary);
-            }
-
-            .PlayerSeekbarTick__line {
-                position: absolute;
-                width: var(--tick-width);
-                margin-left: calc(var(--tick-width) / 2 * -1);
-                border-radius: 2px;
-                opacity: 0.5;
-                height: 100%;
-                background-color: var(--tick-color);
-            }
+        &:hover {
+            --tick-width: 5px;
 
             .PlayerSeekbarTick__info {
-                position: absolute;
-                bottom: 100%;
-                background-color: var(--tick-color);
-                padding: 0.25rem 0.5rem;
-                font-size: 0.8rem;
-                border-radius: 0.25rem;
-                transform: translateX(-50%);
-                white-space: nowrap;
-                color: #fff;
-                margin-bottom: 0.25rem;
-                opacity: 0;
-                visibility: hidden;
-                transition: opacity 150ms ease-in-out;
-                max-width: 20rem;
+                opacity: 1;
+                visibility: visible;
             }
-
-            &:hover {
-                --tick-width: 5px;
-
-                .PlayerSeekbarTick__info {
-                    opacity: 1;
-                    visibility: visible;
-                }
-                .PlayerSeekbarTick__line {
-                    opacity: 1;
-                }
+            .PlayerSeekbarTick__line {
+                opacity: 1;
             }
         }
     }
